// WorkSpace/projects/user/src/app/models/Product/iproduct-read.ts
export interface IProductRead {
  id: number;
  name: string;
  description: string;
  image: string;
  dateTime: Date;
  companyId: number;
  companyName: string;
  companySlogan: string;
  companyEmployees: number;
  companyImage: string;
  companyEmail: string;
  companyPhone: string;
  companyLink: string;
}